Release/v1.9.0

+++ include/WCSimWCAddDarkNoise.hh
@@ -23,16 +23,16 @@ public:
   
 public:
   void AddDarkNoise();
-  void AddDarkNoiseBeforeDigi(WCSimWCDigitsCollection* WCHCPMT, float num1 ,float num2);
-  void FindDarkNoiseRanges(WCSimWCDigitsCollection* WCHCPMT, float width);
+  void AddDarkNoiseBeforeDigi(WCSimWCDigitsCollection* WCHCPMT, double num1 ,double num2);
+  void FindDarkNoiseRanges(WCSimWCDigitsCollection* WCHCPMT, double width);
   //As it inherits from G4VDigitizerModule it needs a digitize class.  Not used
   void Digitize() { }
   void SetDarkRate(double idarkrate){ PMTDarkRate = idarkrate; }
   double GetDarkRate() { return PMTDarkRate; }
   void SetConversion(double iconvrate){ ConvRate = iconvrate; }
   void SetDarkMode(int imode){DarkMode = imode;}
-  void SetDarkHigh(int idarkhigh){DarkHigh = idarkhigh;}
-  void SetDarkLow(int idarklow){DarkLow = idarklow;}
+  void SetDarkHigh(double idarkhigh){DarkHigh = idarkhigh;}
+  void SetDarkLow(double idarklow){DarkLow = idarklow;}
   void SetDarkWindow(int idarkwindow){DarkWindow = idarkwindow;}
   int GetDarkWindow(){return (int)(DarkWindow);}
   void SaveOptionsToOutput(WCSimRootOptions * wcopt, string tag);
@@ -53,8 +53,8 @@ private:
   WCSimDetectorConstruction* myDetector;
   G4String detectorElement;
 
-  std::vector<std::pair<float, float> > ranges;
-  std::vector<std::pair<float, float> > result;
+  std::vector<std::pair<double, double> > ranges;
+  std::vector<std::pair<double, double> > result;
   
 };
